backend/requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
pydantic
uvicorn[standard]
pydantic[email]